LayoutTests/accessibility/ios-simulator/dialog-did-become-focused-expected.txt

Alert Box

Alert text

This tests that accessibilityElementDidBecomeFocused on elements inside a dialog won't set focus onto the dialog.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dialog === document.activeElement is false
PASS successfullyParsed is true

TEST COMPLETE
